pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "genomiq"
version = "1.0.0"
description = "GenomIQ — Scientific Discovery RL Environment"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "requests>=2.31.0",
    "gradio>=4.0.0",
    "plotly>=5.18.0",
    "pandas>=2.1.0",
    "PyYAML>=6.0.1",
]

[project.scripts]
server = "server.app:main"
ui = "gradio_app:demo.launch"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "server*"]
